Se reestructuro enrutamiento, se agrego área de fideicomisos en propiedad, se arego fideicomiso activo a certificado de gravamen, se hace caratula para fideicomisos, se implementa funcion para correccion en inscripciones de propiedad, se hace trait para recuperar propietarios de movimientos registrales anteriores

+++ app/Livewire/Comun/Actores/PropietarioCrear.php
@@ -61,7 +61,7 @@ class PropietarioCrear extends Component
 
         $pu = $pu + (float)$this->porcentaje_usufructo + $pp;
 
-        if((int)$pn > 100 || (int)$pu > 100) throw new ActoresException("La suma de los porcentajes no puede exceder el 100%.");
+        if((float)$pn > 100 || (float)$pu > 100) throw new ActoresException("La suma de los porcentajes no puede exceder el 100%.");
 
     }
